1112
전시용최종빌드

밀 이벤트 연결 안되던거,
바다 이벤트 진짜 바다 타일만 요구하던거 변경

기타 등등 밸패,
물질 1렙 보너스는 이동력 2 이상만 적용되도록

+++ lehoo/Assets/Script/DebugScript.cs
@@ -141,7 +141,7 @@ public class DebugScript : MonoBehaviour
     if (GameManager.Instance.MyGameData.QuestType == QuestType.Cult)
     {
       GameManager.Instance.MyGameData.Quest_Cult_Phase = int.Parse(Cult_Phase.text);
-      GameManager.Instance.MyGameData.Quest_Cult_Progress = int.Parse(Cult_Progress.text);
+      GameManager.Instance.MyGameData.Quest_Cult_Progress = float.Parse(Cult_Progress.text);
 
       UIManager.Instance.SidePanelCultUI.UpdateUI();
     }
